Moving create tests memory function to user_driver_emulated_mock
