+++ vm/devices/user_driver_emulated_mock/Cargo.toml
@@ -11,6 +11,8 @@ anyhow.workspace = true
 chipset_device.workspace = true
 guestmem.workspace = true
 inspect.workspace = true
+memory_range.workspace = true
+page_pool_alloc.workspace = true
 parking_lot.workspace = true
 pci_core.workspace = true
 safeatomic.workspace = true
